Convert MavenVersion to an interface
